Update references: - Updated Microsoft.Data.SqlClient from Version="2.1.7" to Version="5.1.6", which required the following updates:   - Replaced System.Drawing.Common with GeneXus.Drawing.Common.   - Updated "Microsoft.Identity.Client" from Version="4.60.3" to Version="4.61.3". - Continued using System.Drawing.Common for generating graphic reports sent directly to printers, which is supported only on Windows.
