+++ dotnet/src/dotnetcore/GxNetCoreStartup/GxNetCoreStartup.csproj
@@ -11,7 +11,7 @@
 
 	<ItemGroup>
 		<PackageReference Include="Microsoft.AspNetCore.DataProtection.StackExchangeRedis" Version="8.0.3" />
-		<PackageReference Include="Microsoft.Data.SqlClient" Version="2.1.7" />
+		<PackageReference Include="Microsoft.Data.SqlClient" Version="5.1.6" />
 		<PackageReference Include="Microsoft.Extensions.Caching.SqlServer" Version="3.1.3" />
 		<PackageReference Include="Microsoft.Extensions.Caching.StackExchangeRedis" Version="8.0.3" />
 		<PackageReference Include="Microsoft.Identity.Client" Version="4.61.3" />
