+++ dotnet/src/dotnetcore/GxPdfReportsCS/GxPdfReportsCS.csproj
@@ -17,7 +17,7 @@
 	  <EmbeddedResource Include="..\..\dotnetframework\GxPdfReportsCS\sRGB Color Space Profile.icm" Link="sRGB Color Space Profile.icm" />
 	</ItemGroup>
 	<ItemGroup>
-		<PackageReference Include="PdfPig" Version="0.1.8"  PrivateAssets="All"/>
+		<PackageReference Include="PdfPig" Version="0.1.8" PrivateAssets="All" />
 	</ItemGroup>
 
 	<ItemGroup>
@@ -32,8 +32,8 @@
 			<IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
 		</PackageReference>
 		<PackageReference Include="Microsoft.Win32.Registry" Version="4.7.0" />
-		<PackageReference Include="System.Drawing.Common" Version="4.7.2" />
-		<PackageReference Include="System.Text.Encoding.CodePages" Version="4.7.0" />
+		<PackageReference Include="GeneXus.Drawing.Common" Version="2.0.0" />
+		<PackageReference Include="System.Text.Encoding.CodePages" Version="6.0.0" />
 	</ItemGroup>
 
 	<ItemGroup>
